refactor: move evict and download to layer cache
